fix(landcover): declare patchtypes as integer instead of converting at each use
patchtypes is a category (0 soil, 1 urban, 2 wetland, 3 ice, 4 lake), but it
sat inside the real(r8) declaration block of canopy parameters in MOD_Const_LC.
Both source tables it is filled from (patchtypes_igbp / patchtypes_usgs) are
integer parameters, and every consumer is integer too: patchtype(:), pthtype,
and patch_has_pft. The only reason the mismatch stayed invisible is that all
existing uses compare it against an integer literal.

Declare it integer at the source rather than papering over the type at the
call site, which is what the previous commit did with nint(). This also
removes an implicit real-to-integer truncation in
MOD_Initialize.F90: patchtype(ipatch) = patchtypes(m).

Checked every use across the tree: all are integer comparisons, assignments to
integer variables, or size() queries. None treats it as floating point.

+++ mkinidata/MOD_PercentagesPFTReadin.F90
@@ -43,9 +43,7 @@ CONTAINS
 #ifdef SinglePoint
       ! Wetland owns a WFT tile under WETLAND_PFT, so it must not bail out
       ! here: pftfrac would stay unset and the tile would carry no area.
-      ! nint: MOD_Const_LC declares the runtime patchtypes array as real(r8)
-      ! even though it is filled from an integer parameter table.
-      IF (.not. patch_has_pft(nint(patchtypes(SITE_landtype)))) RETURN
+      IF (.not. patch_has_pft(patchtypes(SITE_landtype))) RETURN
 #endif
 
 #ifndef SinglePoint
